feat: 코멘트 작성 api 구현

+++ src/main/java/or/sopt/soptwatcha/repository/KeywordRepository.java
@@ -3,5 +3,5 @@ package or.sopt.soptwatcha.repository;
 import or.sopt.soptwatcha.domain.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordRepository extends JpaRepository<Keyword, Integer> {
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
 }
